Making everything a bit more efficient + use the right types

+++ helpers/privileges.go
@@ -3,6 +3,6 @@ package helpers
 import "github.com/Gigamons/common/consts"
 
 // HasPrivileges if user has those permissions else not!
-func HasPrivileges(p int, u *consts.User) bool {
-	return u.Privileges&int32(p) > 0
+func HasPrivileges(p uint64, u *consts.User) bool {
+	return u.Privileges&p > 0
 }
